requirements.txt
streamlit
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
nltk
sklearn
scikit-learn
openpyxl
